rename dir/file
