+++ algorithms/java/src/myQueue/MyQueue.java
@@ -16,7 +16,7 @@
  *  You may assume that all operations are valid (for example, no pop or peek operations will be called on an empty queue).
  *
  **********************************************************************************/
-package _232_myQueue;
+package myQueue;
 
 import java.util.Stack;
